Refactor out options
