+++ options.h
@@ -1,0 +1,20 @@
+#ifndef OPTIONS_H_
+#define OPTIONS_H_
+
+#include <getopt.h>
+#include <stdio.h>
+
+static const char DEBUG_OPTION[] = "debug";
+static const char EXACT_SEARCH_OPTION[] = "exact";
+static const char HELP_OPTION[] = "help";
+static const char MANAGERS_OPTION[] = "managers";
+
+static struct option long_options[] = {
+    { DEBUG_OPTION, no_argument, NULL, 'd' },
+    { EXACT_SEARCH_OPTION, no_argument, NULL, 'e' },
+    { HELP_OPTION, no_argument, NULL, 'h' },
+    { MANAGERS_OPTION, required_argument, NULL, 'm' },
+    { NULL, 0, NULL, 0 }
+};
+
+#endif
